aes: Enforce alignment of static array
When building on-device unit test binaries, the static array in
test/aes.c were not address-aligned after being linked into the
final binary. This causes "runtest" to emit an exception on device
as the MCU tries to load from the address of the static array which
is unaligned address. Enforce alignment to avoid the error.

BRANCH=nocturne
BUG=b:146059307
TEST=make BOARD=nucleo-h743zi tests -j
     # Flash build/nucleo-h743zi/test-aes.bin
     # Connect to UARD console
     runtest

Change-Id: I91e8523fa0f13bf89baa609cded36e111e466972
Signed-off-by: Yicheng Li <yichengli@chromium.org>