ectool: Fix ectool motionsense info command
Commit 267da3cfc increased the response size of the motionsense "info"
command implicitly by adding a new version of the command.  ectool
explicitly gets the latest version of the command to use, and the kernel
was complaining because it only requested 16 bytes, but the new version
sends back 19.  Hence, update the expected size of the response to
compensate for this.

BUG=b:147472780, b:147720232
BRANCH=firmware-hatch-12672.B
TEST=ectool motionsense info X now returns data instead of failing

Change-Id: Ib297c02f55406b4dc3b8bb0b6b0843165b350995
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2007897
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>