Add `ectool` command for tunneling I2C
This uses matthewb@'s not-yet-upstreamed i2c-pseudo driver
(https://crrev.com/c/1454878) to create a "fake" I2C bus on the system.
Traffic is forwarded to the given I2C port on the EC. This is useful for
running applications that assume a direct connection to their
peripheral.

TEST=Run with an EC that has an I2C master, and try some simple
     operations with its peripherals using `i2ctransfer`.
BUG=chromium:1008568
BRANCH=none

Signed-off-by: Harry Cutts <hcutts@chromium.org>
Change-Id: Idc615d7f3eccac46d0cdc49ca796fe28263eb237