Add some hacks to get long I2C reads working
This patch is not intended for submission; it's too hacky.

The I2C-specific response header only allocates one byte for its size
field, the checksumming fails on the host for large responses (since
the EC is summing all the bytes in the response, but the host only sums
the first `size % 256`). As a hack, if the initial checksum fails, try
again using the expected size of the response.

TEST=Try some long (~900 byte) reads from a Goodix GT7288 over the I2C
     pass-through.
BUG=chromium:1008568
BRANCH=none

Change-Id: I74a5b589f524cdaeeadf02fb62fed6cc5ec3bd66
Signed-off-by: Harry Cutts <hcutts@chromium.org>