puff: always claim AC power is online
powerd logs lots of warnings if it can't see any online power supplies,
which could be the case when the EC comes up without the barrel jack
then boots with it- the host flag never gets updated.

Since we don't have a battery on puff and we don't allow switching power
sources while the system is on (we gate the inactive one), just always
say there is external power (because there is).

BUG=b:147977614
TEST=/sys/class/power_supply/AC/online is now always 1, even when
     running on USB-C.
BRANCH=None

Change-Id: I87345b8a3f07a88f2dccda4b7226bf0b8ee07e2f
Signed-off-by: Peter Marheine <pmarheine@chromium.org>