power/cometlake-discrete: correctly update the SM on forced shutdown
Without this, apshutdown on the EC console turns off the system but it
thinks it's in S5 when it's actually in G3 so it doesn't respond to
power button presses to turn back on. Flag when forcing a shutdown
so we correctly transition into G3.

This also removes eSPI reset handling, which I found to be meaningless
while investigating how other platforms do this. skylake seems to do
that power button dance to force shutdown because it can't directly
control the AP power without also killing the EC, and cometlake copied
that without any apparent need which in turn led to its presence here.

TEST=puff goes directly to G3 on apshutdown
BUG=b:147461413
BRANCH=None

Change-Id: Ic216b5e908b8a77d8d016f60dd8b0de832cf301e
Signed-off-by: Peter Marheine <pmarheine@chromium.org>