Update patch set 1
Patch Set 1:

There's a couple of things here that got my attention:

"it thinks it's in S5"
So I presume this means that the power state is POWER_S5, but how did it get there?
Presumably by detecting the state of the power rails.
So why is it getting it wrong? I can see that POWER_S5G3 common handling calls shutdown_s5_rails()
and then chipset_force_g3(). Is this the more correct path through the state machine?
So the concerns I have here is:
 - getting the state correct is important presumably
 - setting a special flag to correct the state machine seems fraught with risk. If the state machine handling is incorrect, we should correct that it seems.
 - my intuition is that instead of using a special (separate flag), this state should be encoded as part of the state machine.

However, I recognise that these state machines are complex, and I don't claim to have that much familiarity with them. I am not sure if it would be useful for you to walk it through for Sam and I on a whiteboard.

The eSPI reset handling removal could be done in a separate CL.

Patch-set: 1
Reviewer: Gerrit User 1281040 <1281040@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>