pi3usb3x532: Pack i2c port in port_addr
The PI3USB3X532 driver in the EC assumes that all superspeed muxes are
on the same i2c bus, I2C_PORT_USB_MUX.  However, that may not be true
for some boards.  This commit utilizes the MUX_PORT(...) macro to
determine the i2c port to use from the usb_mux table.  The boards that
use this driver have been updated to pack the i2c port in the port_addr
member.  There should be no functional changes to those boards.

BUG=b:147689445
BRANCH=None
TEST=`make -j buildall`

Change-Id: If6460b84a5e39610d658f06a42ca1db0bd4da048
Signed-off-by: Aseda Aboagye <aaboagye@google.com>