PS8805: Do not enable and disable DCI registers
PS8805 is not like PS8751, which needs to rewrite the TCPC reg A0 bit 0
to enable the DCI registers. The DCI registers are always accessible.

Also there is a bug on firmware, like 0x8, 0xC, 0xD, that rewriting the
reg A0 bit 0 to 1 will make the TCPC I2C not accessible.

BRANCH=None
BUG=b:147772854
TEST=Verified on Trogdor, fw 0xC, 0xD, and 0xF, the TCPC I2C is
accessible.

Change-Id: Ie554d2b4022397801423fb3670305bf536b2cc20
Signed-off-by: Wai-Hong Tam <waihong@google.com>