Update patch set 1
Abandoned

not sure what your point was, but as i noted in the commit message, nothing reads COMMIT-QUEUE.ini.  so it has no relation to whether this repo is handled by the CQ.

it's a moot point now anyways as CL:2008088 landed which deleted the file from this branch.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon