Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > Patch Set 10:
> > 
> > This needs to land on master first, then we can cherry-pick it over to octopus firmware branch
> 
> Master's I2C address type has been changed to 7b, but octopus branch still used 8bit addressing, does octopus need these two CLs(crrev.com/c/1704792 and crrev.com/c/1699893)?
> We can change it on master if octopus need change I2C address type to 7b.

We don't need to pull over the 7b addressing to octopus as that would be very difficult to cherry pick at this point.

For this CL, we just need to make a 8b version to land on ToT first, then when we cherry pick it (with the same Change-Id), we will have to modify the code to use 7b addressing.

Patch-set: 10
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>