TCPMv2: Ensure Rd is set in Debug Accessory.SNK
When tcpm_init() runs (ex. during a phy layer reset), it will cache an
Open state for the CC pull.  This can cause Open to be presented on the
CC lines during the next call to set polarity.  The Attached.SNK state
already ensures Rd is set by calling set_cc explicitly, and this change
adds an explicit set to Rd for Debug Accessory.SNK as well.

BRANCH=None
BUG=b:147316570
TEST=loaded on waddledoo, ensured CC lines weren't set to Open when
suzy-q was plugged in

Change-Id: I17ea735632e10c666691c04d56057c57688dcbd6
Signed-off-by: Diana Z <dzigterman@chromium.org>