Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1

Change looks good, but I'd like this change to be tested against the v2 stack.  I'll check with Caveh for instructions on updating the retimer firmware on the USB3 daughterboard to see if that can be used even before the "ectool pdcontrol" is available on the v2 stack.

Patch-set: 3
Reviewer: Gerrit User 1308749 <1308749@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>