npcx7: Add support for ADC threshold interrupts
The Nuvoton NPCX7 supports a threshold event detector for the ADC
channels. There are three threshold event detectors and they can be
mapped to any ADC channel.  The block compares the measure voltage input
to either a programmable threshold or a pair of thresholds and generates
an interrupt if the measured voltage is above or belowe the thresholds.

This commit adds support for this feature.  A new structure is created,
npcx_adc_thresh_t which allows a board to define the ADC channel that
sholud be monitored, the assertion and deassertion thresholds, a
callback that can be run when the interrupt fires and a flag for
whethere or not the interrupt should fire if the measured value is above
or below the threshold.

BUG=b:148169171
BRANCH=None
TEST=Enable on waddledoo, verify that the ADC dual threshold interrupt
functionality works.

Change-Id: I5a3e517207a71b1298865fe36b80cc6298567e9e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>