Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Pardon my question, as this bit me earlier:
> Are you certain this addresses *all* valid exit states from AttachedWait.SRC?
> 
> * Attached.SRC
> * UnorientedDebugAccessory.SRC
> * AudioAccessory
> 
> The third one is equally valid, but I don't see reference to that here.
> Will ping tomorrow, when I have better understanding of original issue.

Take a look at crbug:1046851 which has a good summary of what needs to happen.  This set here is only necessary because a tcpm_init cleared out tour cached Rd value, which the super state had previously set.  We believe there are several tcpm_init calls in the new stack which can be removed, which will in turn remove the need for setting Rd here.

Patch-set: 2