battery: add smart battery as locatable chip
We want to give OS access to all of the smart battery metrics. Allow OS
to get the port and address of the smart battery if present on the
system. This will allow the OS to use i2c tunnel to query the battery
according to the well-known smart battery spec.

BRANCH=none
BUG=none
TEST=verified on octopus that battery info is returned

Change-Id: If8bde4853380e574ecdadfee2e89d4f77b1681da
Signed-off-by: Jett Rink <jettrink@chromium.org>