Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > I have a couple of concerns with this.
> > 1) It allows (presumably) the kernel to read/write any battery/fuel gauge register it wants, regardless of whether or not it's a Smart Battery register, which I don't really think is what we want. I think we really want a new host command that exposes smart battery information in a read-only manner. 
> > 2) Currently, we "close" the TCPC passthru bus after EC and/or AUXFW software sync are completed for security reasons. There may be some battery packs that allow firmware update over this bus, leaving them vulnerable to malicious updates.
> 
> Also there may be other batteries which don't fully comply with SBS, but provide the same information in a different way, so we may want to support those as well.

Thanks for looking over this!

One reason for this more "open i2c" approach is that we don't have to release new EC firmware when we want to collect new smart battery information in the OS. Instead we only have to add kernel code to get the new battery information we need without an EC FW release.

It looks like we are moving more trust to the kernel driver layer to not abuse the interface. However, the i2c read/write passthru command is already available to the kernel, so we aren't opening any more access, we are just making it easier to find the battery (but you could have always talked with the battery).

If open access to the battery is too dangerous, then the EC firmware should lock down writes access on the tunnel (or read/write access), which would apply to all kernel/AP firmware/ectool calls

If the kernel only exposes proper smart battery reads, then userspace can't hurt the battery/EC through that interface. The kernel can already bork your system if it is compromised, right?

Patch-set: 3
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>