TCPMv2: Allow reduced functionality to PD2.0 features
Some TCPCs only work with PD2.0 messages. So a config flag
is added to restrict the TCPMv2 stack to working at PD2.0
revision.

If the config flag is defined, the PD revision is fixed at
PD2.0. In a future CL, the config flag can also be used
to reduce code size by not compiling in PD3.0 features.

BUG=chromium:1021235
BRANCH=none
TEST=make -j buildall
	Manual tests:
	Verified that the stack only works at PD2.0 rev,
	using a twinki and several docks.
Signed-off-by: Sam Hurst <shurst@google.com>

Change-Id: Ic5fc8d24baec3bdecffb4dd54a6aac28d72550cb