volteer: enable minimum power requirement
The TigerLake AP requires in excess of 15 W to power on. Set the minimum
required threshold to boot with no battery to 15.001 W so that PD
negotiation is required before the AP can boot without a battery.

BUG=b:144198315
BRANCH=none
TEST=make buildall
TEST=verify EC powers up when connecting a SuzyQ to a board without
battery and confirm the board stays in G3. Verify the board boots
normally when connecting a 45W PD supply.

Change-Id: I51e01d31d014432a15743e6f407ba157be9bbff6
Signed-off-by: Keith Short <keithshort@chromium.org>