common/system: Unify ec_current_image and system_image_copy_t
"enum ec_current_image" is exposed in ec_commands.h (and used by non-EC
code, such as biod).

We also have an "enum system_image_copy_t" that is the exact same thing
(though has a few more definitions).

A followup commit adds the "sysinfo" host command, so we want to be able
to expose all the potential image variants. Rather than maintain two
enums that can potentially get out of sync, unify the code to use a
single enum. We choose to keep the "enum ec_current_image", since
external code depends on it.

To verify that this change results in no changes to the generated
binaries, we use the script in https://crrev.com/c/2001130:

./util/compare_build.sh --board all

BRANCH=none
BUG=b:146447208
TEST=./util/compare_build.sh --board=all

Change-Id: I13776bc3fd6e6ad635980476a35571c52b1767ac
Signed-off-by: Tom Hughes <tomhughes@chromium.org>