Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Daisuke seemed to indicate in this CL that CHIPSET_RESUME was insufficient. https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1768917

So as I read the comments, it seems for Fizz that the RESUME hook is unreliable, but I can't see a separate bug for that. Perhaps that should be fixed separately.
The issue here is that CHIPSET_RESET is called for both startup and shutdown (I have confirmed this by observation), whereas the CL referenced (and comments) seem to assume that CHIPSET_RESET is only called on startup.
I suspect that there is a bit of a timing race here - what I saw on Puff was CHIPSET_SHUTDOWN followed by CHIPSET_RESET - perhaps on Fizz the order was reversed.

In any case, I don't know what the definitive correct thing to do here - is it the contract that CHIPSET_RESET should only be invoked when the processing is starting up? In which case the bug is that it is being invoked also when the processor is shutting down.
I don't know the expectations for these hooks - are they meant to be reliable and definitive signals?

Patch-set: 2