palkia: add keyboard customization
As the Palkia isn't using chromeos keyboard, we need to
remapping the keyboard matrix for the key event. The
new config option CONFIG_KEYBOARD_CUSTOMIZATION is added
to customize the keyboard matrix in the board layer.
Most of the keys have been remapped except F11-F16, Fn and
win key, we can remap it later.

BUG=b:148034320
BRANCH=firmware-hatch-12672.B
TEST=esc+refresh+power
     alt+vol up+r
     evtest for all of the keys

Change-Id: Ibb02e3a7f5e022eeca36865e8561393fc326f2af
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>