Makefile: enforce target name in generated dependencies
There seems to be some odd interaction between ccache version 3.7.6
and the set of command line options passed to gcc by the EC makefile,
as a result the generated dependency files are wrong, the target file
name is missing the path.

The -MT command line option makes sure that the correct target file
name is generated. Had to make similar changes in
../../third_party/{cryptoc,tpm2} Makefiles.

No need to change extra/usb_updater/Makefile as it puts .o files in
the same directory where .c files are.

BRANCH=all
BUG=b:148943341
TEST=verified that relevant object files are rebuilt when an .h file
     is touched.

     Also, with companion changes in cryptoc and tpm2 trees verified
     that all generated my 'make buildall' .d files have proper target
     values (including path), apart froom files in extra/usb_udater,
     which place .o files in the same directory with .c files.

Change-Id: I22dfad94c112582230a2b6b60289e029a382d822
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>