Update patch set 2
Patch Set 2:

+1 on your comments, will work with Sam and you tomorrow to understand code better. I'm a noob.

> Maybe I'm too focused on the solution I have in mind, but this still doesn't seem like the correct fix in my opinion. I think the problem is we're setting the SuperSpeed muxes to some "default" state on every chipset state transition.
> 
> However, I think that they should only be configured upon attach, detach, and DP configure. Maybe also disconnect the mux when we're not using any of the lines at all (such as when the system is shutdown completely).

We should also isolate pins when not resolved in state. One of USB-IF "Functional Type-C" tests checks if we are properly removing terms when not resolved SNK or SRC.

> Also, by removing that event notification, do we still exit DP mode when shutting down?

Let me test and check back, excellent catch. Implications of not exiting are dongle not working the next boot, if left connected.

That would be a degenerate case of "Double-AltMode-Entry" -- which many dongles really dislike.

Patch-set: 2
Reviewer: Gerrit User 1269501 <1269501@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>