Update patch set 1
Patch Set 1:

I still think that repeating the same comment multiple times does not improve things but adds noise to the code.

Also it looks odd that out of multiple gcc command line options only one is commented, to be consistent one would need to describe all of them, in all places.

But if someone thinks this is a worthy addition - it's fine with me.

Patch-set: 1
CC: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>