efs2: update parsing response logic
Update parsing logic to treat response as a stream of bytes that is
interpreted as the response structure so parsing code can be independent
of what is in the response from cr50

Also, retry failing CR50/EC communication on any error up to fives
times in case there was a problem transmitting the data across the wire.

See CL:2071059 for cr50 update as well.

BRANCH=none
BUG=none
TEST=builds on puff when I manually make it EFS2

Change-Id: I95f373c40d9a945de056217bc8f519a6a54463a1
Signed-off-by: Jett Rink <jettrink@chromium.org>