syv682x: Enable smart discharge mode
The SYV682A device provides a smart discharge mode that automatically
discharges VBUS and VCHG when the channel is shutdown. This works better
than the manual dicharge mode which must be explicitly cleared after the
discharge completes.

This change also adds BUSY detection to the driver.

BUG=b:148870148, b:148467221
BRANCH=none
TEST=make buildall
TEST=Verify connection of SRC devices start battery charging, including
DRP devices.
TEST=Verify SNK devices are detected and reported by the kernel
TEST=Measture VBUS discharge times of 20 ms when disconnecting both SNK
and SRC devices.

Change-Id: I54a5367924f3ae5adc473cb4b0a11789476e7bc3
Signed-off-by: Keith Short <keithshort@chromium.org>