libcryptoc: Remove default unaligned support
The SUPPORT_UNALIGNED had been enabled for all chips.
This is not correct. An unaligned access on most chips will
result in an immediate exception.
This has not been exposed due to the current sizing and offsets used
in libcryptoc's current implementation, but this may not hold
for new changes to libcryptoc.

It appears that Cr50 first introduced the SUPPORT_UNALIGNED flag.
This usage of libcryptoc was then copied for general usage.
Cr50 now uses this generalized libcryptoc recipe, but Cr50
already adds the -DSUPPORT_UNALIGNED=1 CFLAG, thus libcryptoc
is compiled with the support.

BRANCH=none
BUG=b:143090229
TEST=rm -rf build/
     make BOARD=cr50 V=1 build/cr50/cryptoc/libcryptoc.a
     # Ensure that SUPPORT_UNALIGNED=1 is still set
     # during compilation of objects.
     # You should still see "-DSUPPORT_UNALIGNED=1".
TEST=rm -rf build/
     make BOARD=nocturne_fp V=1 build/nocturne_fp/cryptoc/libcryptoc.a
     # Ensure that SUPPORT_UNALIGNED=1 is NOT still set
     # during compilation of objects
TEST=rm -rf build/
     make BOARD=hatch_fp V=1 build/hatch_fp/cryptoc/libcryptoc.a
     # Ensure that SUPPORT_UNALIGNED=1 is NOT still set
     # during compilation of objects

Change-Id: I7572b418a0330bd19c1ab4c8f4a64fd3eece89a0
Signed-off-by: Craig Hesling <hesling@chromium.org>