kukui: inform AP DP status by USB_MUX event
This CL notify the usb mux status to AP no matter the it5205 exists or
not.

Some revs of the krane and kodama replace virtual mux with a real mux.
However, this will stop informing the AP usb mux status, but it6505
still need the mux information to control the regulators.

BUG=b:149181702
BRANCH=kukui
TEST=plug hub w/ DP port, and see exton state DP=1 by
     'cat /sys/class/extcon/extcon0/state'

Change-Id: I86c0998340c935b16b02ec0c3edf0484f1843ebf
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>