Update patch set 1
Patch Set 1:

yupingso@: Is there anything we can trim from kakadu to free up more RO space?

This change goes over the RO flash space for kakadu, but I think it would be a useful command to have for all boards, so I'm hesitant to disable it for kakadu if we don't have to.

Patch-set: 1
CC: Gerrit User 1353551 <1353551@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>