Update patch set 2
Patch Set 2: Verified-1

i'm not seeing VBUS turn on with this patch on the USB3 board.
could you double check with this setup:

- charging from C0
- battery attached (full in my case)
- USB3 DB

boot to linux with nothing connected on the USB3 board.  then,
insert a USB-C to A dongle with an LED or similar load to check
if VBUS comes on.

Patch-set: 2
Reviewer: Gerrit User 1190764 <1190764@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=-1