Update patch set 4
Patch Set 4:

> Once again, i2c is not meant to be discovered at run time: the bus topology is supposed to be well defined at build time.
1. May I know where is the standard for this rule? Thanks.
2. As I know, in the uni-build image the ACPI defines all possible components on I2C buses then kernel driver will get chance to probe it failed by I2C command. So it is the case in ChromeOS now.
3. In octopus EC, if the sub-board is for 1 USBA only not 1USBA and 1USBC then our EC codes depend on the fail of probing to TCPC chip then stop a PD task. This is a second case.

> motion_sense_switch_sensor_rate (deferred function) <- motion_sense_resume (HOOK_CHIPSET_RESUME,) <- motion_sense_startup (HOOK_INIT)

Looks like sensor initialization is done in motion_sense_switch_sensor_rate as a deferred function now if sensor state is not SENSOR_INITIALIZED and SENSOR_ACTIVE. Could we hook sensor probe here? For example, if motion_sensor_count is
    - -1: (wait_for_probing) then calling sensor_probe like function (or call all init() of candidate sensors one by one) to check sensor and update motion_sensors array.
    - 0: no sensors
    - more then 0: skip probing

In this case, probe and init will be in the same hook call and deferred function of motion_sense_switch_sensor_rate?

If it sounds reasonable then I can try it?

Patch-set: 4