Update patch set 19
Patch Set 19:

I don't think this generalization is necessary, it adds a lot of code which is unlikely to be useful.

Also, what happens with the EC console output which could be sent interleaved with EC CR50 communications? Is it held off on the EC, or is it OK to drop it and not pass to USB if EC CR50 comms are active?

Patch-set: 19
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>