syv682x: Don't turn off VBUS when disabling sink mode
When the last charger is removed from the system, boards call
ppc_vbus_sink_enable() to disable sink mode on all PD ports. The SYV682
doesn't support turning of VBUS and VCHG independently, so do nothing if
if the SYV682 is currently the VBUS source.

BUG=b:149428774
BRANCH=none
TEST=make buildall
TEST=On volteer, connect a charger to C0, connect a SNK device to C1.
Disconnect charger and C0 and verify VBUS stays up on C1.

Change-Id: I1747b3b19f584b0069134c520bd0ef0453a0a06a
Signed-off-by: Keith Short <keithshort@chromium.org>