chgramp: Don't ramp DTS suppliers above advertisement
We shouldn't ramp above the current limit that the DTS specifies; the
USB Type-C spec clearly shows what the limit is.  The existing code was
allowing us to ramp up to 2.4A when a SuzyQable was connected to a
chromebook.  This was causing issues on various devices/workstations.

If a board needs more power than SuzyQable can provide, they ought to
use a ServoV4 with a charger connected instead.

BUG=chromium:770296,b:142033952,b:144198315
BRANCH=hatch,nami,octopus
TEST=Build and flash kohaku, plug in SuzyQable, verify that DUT draws no
more than 1500mA.
TEST=Repeat above test with a ServoV4 with 60W adapter plugged in,
verify DUT can pull ~60W.

Change-Id: Ife9f8d99862d61cf7d18722fdd6e19331a47e301
Signed-off-by: Aseda Aboagye <aaboagye@google.com>