nct38xx: DRP connection cleanup
Don't set the polarity behind the back of the PD stack.
Just clear the DRP and leave the CC lines so they look
just as we found them.  This will allow TRY to work
and we will no go OPEN because we set the CC lines to
something that was not expected.

BUG=b:149415919
BRANCH=none
TEST=verify USB-C connections are working

Change-Id: I766514bd46922000ea8916d61d00265e7e5e4fd4
Signed-off-by: Denis Brockus <dbrockus@chromium.org>