Update patch set 2
Patch Set 2:

I guess it is a bit too late to argue, but come to think of it, I am not sure this all is worth the effort: This function is called very seldom, this optimization required to alter the third party library, and broke 'make buildall' along the way.

I would rather revert the 0be7f8dcc9 change  in tpm2 library completely and not worry about this, but will defer to other reviewers.

Patch-set: 2
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>