tcpci: Don't set 1 CC to open based on polarity
This commit fixes an issue where the TCPCI driver was setting 1 CC line
to open when the polarity was known.  This was causing issues where a
charger would only work in one orientation without a battery.  Ideally,
this behaviour should be done a bit higher up in the stack whenever we
are sourcing VCONN.

BUG=b:149496312,chromium:149496312
BRANCH=hatch
TEST=Build and flash waddledoo, remove the battery, plug in a PD
charger, verify that the EC does not brown out.

Change-Id: Idbcb915c0bbd867cfafc1e32c0f6534b363a867b
Signed-off-by: Aseda Aboagye <aaboagye@google.com>