Update patch set 11
Patch Set 11:

> But I don't know what is the error relation with CTUnattached.SNK? I hook to unattach.SRC/SNK state not CTunattach.VPD/SNK.

I don't know what this hook is doing in the ITE driver, as there is no CL associated with this one. I also don't see a bug number associated with this change.

Depending on what this "hook" is doing (setting VBUS discharge? turning on/off response modes?) there are other states which may be affected. CT-VPD is one of those states.

We do not know what other people may misinterpret this "HOOK_USB_PD_DISCONNECT" for in the future, which will then introduce strange bugs when they use it inappropriately. All boards from now until the future will get this change, since it is in Master.

I am not a coder, but I believe it might be wiser to approach this from another method that is more generally applicable. Such as splitting the hook into two explicit, easily understood calls: "UnattachedSRC-Entry" and "UnattachedSNK-Entry". This is clear and straightforward.

I would appreciate the insight of more experienced ChromeOS members.

Patch-set: 11