usbc: remove tri-state polarity
TCPCIr2 had an issue with setting CC coming out of DRP
that if the polarity was not retained that the connection
dropped back to OPEN. Unfortunately this change broke
many of the other TCPCI implementations.  I am working
on a different method of dealing with coming out of DRP
and this is no longer needed.

BUG=none
BRANCH=none
TEST=verify USB-C is working

Change-Id: Ifa8f26d417df2f5d5f41a23fbf7e6f9129031e94
Signed-off-by: Denis Brockus <dbrockus@chromium.org>