syv682x: Create interrupt handler
Add interrupt handler to SYV682. All alerts are clear on read, so
interrupts must be checked regardless of why the register is being read.
It is also very spammy so flags are used to trigger only on rising edges
of alerts.

BUG=b:149531621
BRANCH=none
TEST=make buildall
TEST=On Volteer Connect a passive USB C-to-A adapter to the USB4 DB.
Short the VBUS and ground pins; make sure the OC event is seen by the EC
with the C1 overcurrent console message.

Change-Id: I61c1a6b962e60d813adab92fcbd21fbdf838f188
Signed-off-by: Eric Herrmann <eherrmann@chromium.org>