syv682x: Create interrupt handler
Add interrupt handler to SYV682. The 4 interrupts we take care of here
are:
1. HV OCP
2. 5V OCP
3. Thermal Shutdown
4. FRS detected

The first 3 are handled as overcurrent events, while the last calls the
frs event in the PD policy engine.

BUG=b:149531621
BRANCH=none
TEST=make buildall
TEST=On Volteer Connect a passive USB C-to-A adapter to the USB4 DB.
Short the VBUS and ground pins; make sure the OC event is seen by the EC
with the C1 overcurrent console message.

Change-Id: I61c1a6b962e60d813adab92fcbd21fbdf838f188
Signed-off-by: Eric Herrmann <eherrmann@chromium.org>