TCPMv2: Fix GET_SNK_CAP response processing
After GET_SINK_CAP message, we need to wait for a DATA_SINK_CAP
message. The port partner should not respond with anything other than
DATA_SINK_CAP, so the exit condition is only a timeout.

Also, DPM requests were hidden behind a check that we are in a READY
state which doesn't allow the get_sink state to be queued up.

BUG=b:148144711
TEST=make buildall -j
BRANCH=none

Change-Id: I9db09283eca20231d7b1f514485a6fcd90f9335f
Signed-off-by: Eric Herrmann <eherrmann@chromium.org>