TCPMv2: Don't check RAM flag for VCONN
If the EC resets, reality might not match up with our SRAM saved flags.
Previously, this code was gating modifying VCONN based upon the SRAM
flag.  However, it's possible that VCONN is still being sourced but our
flag indicates that it's not.  A call to disable VCONN would be ignored
due to thinking that it's already off.

This commit simply attempts to modify VCONN to the value that was
requested.  There's no harm in turning on VCONN if it's already on, or
turning it off when it's already off, therefore just set it
unconditionally.

BUG=b:149581333
BRANCH=None
TEST=Build and flash waddledoo, boot to S0, plug in a sink, reset the EC
while removing the sink, verify that VCONN is no longer sourced.

Change-Id: Id7ee84c077e05fdfc6fa3096cb92fedcfff3a3f3
Signed-off-by: Aseda Aboagye <aaboagye@google.com>