TCPMv2: Disable VBUS whenever we're a sink
We enter the Rd super state whenever we're a sink.  The Type-C spec says
that in the Unattached.SNK state, the port shall not drive VBUS or
VCONN.  In this state, we weren't driving VCONN, but we weren't
explicitly not driving VBUS.

BUG=b:149581333
BRANCH=None
TEST=Build and flash waddledoo, source to a sink, EC reset while
removing sink, verify that VBUS is no longer sourced.

Change-Id: Id0ba138f97aebdada7a8aca22942e3fd4464624e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>