raa489000: Add workaround for VBUS auto discharge
The Renesas RAA489000 has an issue where GoodCRC messages are not
emitted unless the VBUS Auto Discharge Disconnect setting is enabled.
This needs to occur after determining orientation.

However, there appears to be another issue where the device will not DRP
toggle while that VBUS Auto Discharge Disconnect setting is set.
Therefore, this commit adds a workound to remove that setting when a
device is disconnected in order to allouw DRP toggling to work.

BUG=b:149337338
BRANCH=None
TEST=Build and flash waddledoo, boot to S0, verify DUT DRP toggles.
Plug in a PD port partner, verify that PD communication works.

Change-Id: Ieb79392f3cd7c1939bb6fb3de0d6785cec0b24c4
Signed-off-by: Aseda Aboagye <aaboagye@google.com>