Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> Yes, we hook to do "exiting BIST mode" and "setting plug-in ISR" in tcpm it83xx.c driver. It's in separately CL "https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1170718" and "https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1333207" for TCPMv1 and also needed for TCPMv2.
> 
> Ok, I will change the hook label in another CL.
> 
> No, it will not repeatedly trigger "HOOK_USB_PD_DISCONNECT" hooks when DRP toggle, because there is a if condition outside of hook "if (get_last_state_tc(port) != TC_UNATTACHED_SRC)" or "if (get_last_state_tc(port) != TC_UNATTACHED_SNK)". The hook will be called when the state transit to unattach.SRC/SNK from any state but except unattach.SNK/SRC.

We do need a way for the TCPMv2 to notify other tasks of a disconnect event. This seems perfectly reasonable. Maybe we change how the TCPMv2 calls that disconnect event, but we should not couple other tasks to the intricate knowledge of how the TCPMv2 states are implmeneted. PD_DISCONNECT is a thing that can be understood by external tasks.

Patch-set: 11