Volteer: Add support for FRS to C1
Add support to C1, assign the gpio to put the PPC in FRS mode and add
the config options.

The TCPMv2 stack needs to be enabled to enable FRS. That is tracked in
crrev.com/c/2028355

BUG=b:148144711
TEST=make buildall
TEST=Confirm that if the TCPMv2 stack is turned on, FRS occurs as
expected. The FRS_EN gpio should be set after get_sink_cap if using an
FRS-capable device.
BRANCH=none

Change-Id: I7a599f1c350529d910a331f1ebc78ab41ff44a24
Signed-off-by: Eric Herrmann <eherrmann@chromium.org>