common/flash: Add validation to persist_state size.
Adds a check to validate the size of the persist_state
to verify that it fits within the space allocated.

BUG=b:149509270
TEST=Verified that when the sizeof(persist_state) less than
or equal to CONFIG_FW_PSTATE_SIZE that no change occurs.
Flash space remains the same and read / writes are functional.
Verified that when the persist_state is enlarged, the
BUILD_ASSERT check fails giving a compile time error.
Branch=none

Change-Id: Ia31329ec7817cef71e3835c94028951eaf71fd80