test: Add console channel for tests
This adds a dedicated console channel for test results
and provides test_cprint*/puts macros for tests to use.
This lays the groundwork for test handlers to be able to
distinguish between test results/logs and runtime output.

Note, this channel only exists in test builds and thus
does not change production builds.

BRANCH=nocturne,hatch
BUG=none
TEST=Deploy an on-device test to nucleo and run "runtest"
TEST=Deploy an on-device test to nucleo and run "chan"
TEST=Deploy nucleo-h743zi to nucleo and run "chan" (no test channel)

Change-Id: Idc8029281c53a10479744fa518e521a4e09f45f6
Signed-off-by: Craig Hesling <hesling@chromium.org>