common/flash: Adds validation to setting serial
Adds a simple validation to the serial number to verify it fits
within the buffer allocated to it. If a serial number is passed
that is too long, it will no longer be silently truncated, instead
an invalid argument will show up on the console. Also performed a
small bit of cleanup on the function to eliminate extra loops
and replace them with memset and memcpy functions.

BUG=b:149775650
TEST=Verified that we can set and load serial numbers correctly.
On ServoV4 where CONFIG_SERIALNO_LEN = 28:
Verified that serial numbers up to 27 characters can be set and loaded.
Verified 28 characters and longer return error codes.

Change-Id: I9b4fb269a59023001a0fa8c1e8fc810db320fb8f