waddledoo: Set CMPOUT as inverted to enable 5V
The charger has a general comparator which is enabled by default.  The
output of this comparator enables the sub board power rails.  This
commit adds a mechanism to turn on the 5V on the sub board tied to when
the system decides to enable 5V.

BUG=b:149794574
BRANCH=None
TEST=Build and flash waddledoo, boot to S0, plug in a sink, verify that
VCONN is sourced.

Change-Id: I05384bdbdde4ccc033c2057b3c45fcaeae4589cc
Signed-off-by: Aseda Aboagye <aaboagye@google.com>