Update patch set 2
Patch Set 2: Code-Review-2

I'd like to hold off on this until we settle on our requirements. From what I can gather we keep bolting things on to fix issues that crop up. However we should be maintaining a requirements doc that we can then implement. I need to run to another meeting, and I apologize for the -2, but I think we're approaching development in this area in a non-ideal way. Vijay, that's not your fault. Please be patient while we collectively get our ducks in a row.

Patch-set: 2
Reviewer: Gerrit User 1002369 <1002369@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2