common/keyboard: Start sending action codes by default (aka Vivaldi)
Vivaldi is a new keyboard that allows individual boards to have
additional or different top row keys and/or to reorder those keys.
As part of enabling vivaldi, another feature that needs to be enabled
is to start sending action codes instead of function codes for the top
row of the keyboard.

go/vivaldi-prd
go/vivaldi-design
go/vivaldi-fw-design

With this patch:
 * All board begin to send action scancodes for the top row when top row
   keys are pressed.
 * EC responds to an EC command from the coreboot, that asks for the top
   row layout for the keyboard. Coreboot uses this to expose
   corresponding keycodes to the kernel.
 * Individual boards can customize the top row layout if needed.

BUG=b:146501925
TEST=Check on Jinlon the (new) expected scancodes are output from EC.
BRANCH=firmware-hatch-12672.B

Signed-off-by: Rajat Jain <rajatja@google.com>
Change-Id: I1f45ce6eee138a984f8d4caef1ec76c9538dd30b