common/keyboard_vivaldi: New keyboard framework for custom layout
Vivaldi is a new keyboard that allows individual boards to have
additional or different top row keys and/or to reorder those keys.
The primary agenda of vivaldi is to allow customization of the top
row keys. However, as a secondary objective, it also allows coreboot
to send a keymap to the kernel, for only the keys that are actually
present on the keyboard.

As part of enabling vivaldi, another feature that get enabled
is to start sending action codes instead of function codes for the top
row of the keyboard.

go/vivaldi-prd
go/vivaldi-design
go/vivaldi-fw-design

With this patch, things remain unchanged for any boards that do not provide
board_set_vivaldi_keybd_config(). For boards that do provide it:

 * EC begins to send action scancodes instead of function scancodes for the
   top row when top row keys are pressed.
 * EC responds to an EC command from the coreboot, that asks for the
   layout for the keyboard. Coreboot uses this to expose corresponding
   keycodes to the kernel.

BUG=b:146501925
TEST=Check on Jinlon the (new) expected scancodes are output from EC.
BRANCH=firmware-hatch-12672.B

Signed-off-by: Rajat Jain <rajatja@google.com>
Change-Id: I1f45ce6eee138a984f8d4caef1ec76c9538dd30b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2135566