Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 1) There isn't a standard rule, but b/144811957#comment19 is another example of how i2c-based probing can cause problems
> > 
> > 3) about the sub-board: using the probe failed results instead of using the SKU information ended up causing us problems that we had to fix later. If we had instead based the decision on SKU (via FW_config) we would have avoided later problems when use introduces the LTE subboard.
> 
> Thanks Jett to provide more reference about it. But 
>     1. i2cdetect is common linux tool to really scan the i2c bus so there are really use cases on detection of i2c bus and of course there is a warning on it as you mentioned above.
>     2. I agree that 'SCAN' unknown devices on the i2c bus is dangerous but in this case, there are finite set of i2c clients might be in the bus so we definitely can verify whether any of clients will pull down the SDA / mess the bus and issue limited address to probe. As a result, the risk is under control actually.
>     3. For bad clients messed the bus, I believe that is the issue of this client so we can ask vendor to improve and set this as the AVL requirement.
> 
> If conditions of 2 and 3 above are locked then what else risk we might hit?

Sorry for the delayed response. Can we move this discussion to the bug (b/143440758)?
I will say that i2c probing is definitely technically possible, but there is empirical evidence that is can fail in certain scenarios, and since 1st-class probing isn't something the bus supports since does seems like encoding the hardware topology in CBI makes a lot of sense here.

For octopus specifically, maybe we could introduce the new fw_config field and use that to inform the firmware of which driver to use, then we wouldn't have to keep update the firmware image when we release new IDs that did or did not use the second-sourced version of the the sensors.

Patch-set: 4